+++ Homework_1/split_logic.c
@@ -5,7 +5,7 @@
 #include <stdbool.h>
 #include <stddef.h>
 
-void bubble_sort(waybill_structure *waybill, int num) {
+void bubble_sort(waybill_structure *waybill, unsigned int num) {
 
     if (waybill) {
 
@@ -30,7 +30,7 @@ void bubble_sort(waybill_structure *waybill, int num) {
 }
 
 int waybill_split(waybill_structure *waybill, waybill_structure *waybill_new_1,
-                  waybill_structure *waybill_new_2, int num) {
+                  waybill_structure *waybill_new_2, unsigned int num) {
 
     if ((waybill) && (waybill_new_1) && (waybill_new_2)) {
 
@@ -40,11 +40,11 @@ int waybill_split(waybill_structure *waybill, waybill_structure *waybill_new_1,
         waybill_new_2[0] = waybill[1];
 
         // 2 первых самых больших значение записываем в 2 накладные
-        int num_1 = 1, num_2 = 1;
+        unsigned int num_1 = 1, num_2 = 1;
         double sum_price_1 = waybill_new_1[0].price * waybill_new_1[0].num;
         double sum_price_2 = waybill_new_2[0].price * waybill_new_2[0].num;
 
-        for (int i = 2; i < num; ++i) {
+        for (size_t i = 2; i < num; ++i) {
 
             if (sum_price_1 <= sum_price_2) {
                 waybill_new_1[num_1++] = waybill[i];
